backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
sqlalchemy==2.0.28
psycopg[binary]
pgvector==0.2.5
alembic==1.13.1
pydantic==2.6.3
pydantic-settings==2.2.1
email-validator
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
boto3==1.34.59
botocore==1.34.59
pillow==10.2.0
numpy==1.26.4
python-dotenv==1.0.1
slowapi==0.1.9
httpx==0.27.0
aiofiles==23.2.1
